configurable ca certs configmap for CephObjectStore
The CephObjectStore deployment uses tls to connect to
keystone. This can involve a selfsigned ca certificate.
Currently it is not possible to configure the ca certificate
files used to verify the tls certificates.

This commit adds a new spec entry "caConfigMapName" to
provide a name of a config map. This config map contains
a ca-bundle.crt file (e.g. provided by trust-manager)
which contains the ca certificates to use on verifying a
tls certificate. The config map is included as a volumeMount
in the CephObjectStore deployment.

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>