disable s3 if the swift prefix is set to /
Currently setting the swift prefix to a forward slash
does not disable the s3 api.

In the documentation however there is a red box saying that
the s3 api must be disabled when using a forward slash
as the swift prefix
(https://docs.ceph.com/en/octopus/radosgw/config-ref/#swift-settings)
This commit adds a check on the prefix being a forward
slash and disables s3 if that is the case.

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>